requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
certifi==2023.5.7
charset-normalizer==3.1.0
Django==4.2.2
django-filter==23.2
djangorestframework==3.14.0
idna==3.4
importlib-metadata==6.7.0
Markdown==3.4.3
pytz==2023.3
requests==2.31.0
sqlparse==0.4.4
typing_extensions==4.6.3
urllib3==1.26.6
zipp==3.15.0
